+++ packages/toolkit/src/lib/use-instill-form/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./BooleanField";
+export * from "./CredentialTextField";
 export * from "./OneOfConditionField";
 export * from "./SingleSelectField";
 export * from "./TextAreaField";